New autocompletion and omnisharp settings
We also remove unused plugins.

Added dispatch for building from within vim if you want that. `m<CR>` in
normal mode kicks of a build. I set the makeprg to be `dotnet build` for
cs files. Any errors in the build will be shown in the quickfix window.
The `errorformat` that has been set up for cs files makes it possible to
navigate to errors. For running both build and all tests, hit
backtick-enter. Unfortunately the errorformat does not support jumping
to failing tests right from the quickfix window.

Added vim-test for running tests from within vim. `:TestNearest` is
nice.

Also set omnisharp to use stdio server which works really well. Also
added ultisnips back to mix and let omnisharp get snippet completion
results. What that means is that we get the method signature as a
snippet that we can run through with `<c-n>` / `<c-p>` after having
accepted a completion from the pum with `<c-y>`.