Try out goyo once again
Figured out a few things in my setup that trips up goyo in a way that
makes it display linenumbers all over the place.

In order for that not to happen I have to delete the `autocmd`s in my
`linenumbers` `augroup` and that neeeds to happen before anything
Goyo-related is kicked into gear; That happens in `goyo_start()`.

In `goyo_leave()` we disable limelight and then simply source the vimrc
to restore the settings we had before. The `silent!` part is because
otherwise Vim will complain about redefining a function that's being
used (`goyo_leave`) but that's not something we care about.